+++ integrations/actix-web/src/lib.rs
@@ -3,8 +3,10 @@
 #![allow(clippy::upper_case_acronyms)]
 #![warn(missing_docs)]
 
+mod handler;
 mod request;
 mod subscription;
 
+pub use handler::GraphQL;
 pub use request::{GraphQLBatchRequest, GraphQLRequest, GraphQLResponse};
 pub use subscription::GraphQLSubscription;