+++ src/http/mod.rs
@@ -5,6 +5,7 @@ mod graphiql_source;
 #[cfg(feature = "graphiql")]
 mod graphiql_v2_source;
 mod multipart;
+mod multipart_subscribe;
 #[cfg(feature = "playground")]
 mod playground_source;
 mod websocket;
@@ -18,6 +19,7 @@ pub use graphiql_source::graphiql_source;
 pub use graphiql_v2_source::{Credentials, GraphiQLSource};
 use mime;
 pub use multipart::MultipartOptions;
+pub use multipart_subscribe::{create_multipart_mixed_stream, is_accept_multipart_mixed};
 #[cfg(feature = "playground")]
 pub use playground_source::{playground_source, GraphQLPlaygroundConfig};
 use serde::Deserialize;