+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [6.0.4] 2023-08-18
+
+- Parse "repeatable" in directive definitions. [#1336](https://github.com/async-graphql/async-graphql/pull/1336)
+- add support `multipart/mixed` request. [#1348](https://github.com/async-graphql/async-graphql/issues/1348)
+- async-graphql-actix-web: add `GraphQL` handler.
+- async-graphql-axum: add `GraphQL` service.
+
 # [6.0.3] 2023-08-15
 
 - dynamic: fix the error that some methods of `XXXAccessor` return reference lifetimes that are smaller than expected.
@@ -64,7 +71,7 @@ pub fn lowercase() -> impl CustomDirective {
 
 # [5.0.8] 2023-05-09
 
-- Improve documentation about DataloaderImprove documentation about Dataloader [#1282](https://github.com/async-graphql/async-graphql/pull/1282)
+- Improve documentation on Dataloader [#1282](https://github.com/async-graphql/async-graphql/pull/1282)
 - Prevent recursive input type checking from hitting stack overflow [#1284](https://github.com/async-graphql/async-graphql/pull/1284)
 - update MSRV to `1.65.0`